artifacts/batch_image_processor_full_review_bundle/01_project/project_dir/requirements.txt
Flask==3.1.2
Pillow==11.1.0
